+++ MLEM.Ui/Style/UiStyle.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using MLEM.Font;
 using MLEM.Formatting;
+using MLEM.Formatting.Codes;
 using MLEM.Misc;
 using MLEM.Textures;
 using MLEM.Ui.Elements;
@@ -207,6 +208,11 @@ namespace MLEM.Ui.Style {
         /// Note that this sound is only played if the callbacks have any subscribers.
         /// </summary>
         public SoundEffectInfo ActionSound;
+        /// <summary>
+        /// The color that a <see cref="Paragraph"/>'s <see cref="Paragraph.Link"/> codes should have.
+        /// This value is passed to <see cref="LinkCode"/>.
+        /// </summary>
+        public Color? LinkColor;
         /// <summary>
         /// A set of additional fonts that can be used for the <c>&lt;f FontName&gt;</c> formatting code
         /// </summary>